+++ Source/qol/common.cpp
@@ -14,14 +14,14 @@
 
 namespace devilution {
 
-char *PrintWithSeparator(char *out, long long n)
+char *PrintWithSeparator(char *out, int64_t n)
 {
 	if (n < 1000) {
-		return out + sprintf(out, "%lld", n);
+		return out + sprintf(out, "%ld", n);
 	}
 
 	char *append = PrintWithSeparator(out, n / 1000);
-	return append + sprintf(append, ",%03lld", n % 1000);
+	return append + sprintf(append, ",%03ld", n % 1000);
 }
 
 void FreeQol()
